requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-snappy
numpy
pyyaml
tzdata
nano
vim
gcc
libsnappy-dev
pyarrow
fastparquet